+++ ayrton/__init__.py
@@ -20,11 +20,10 @@
 import os
 import sys
 import importlib
-import builtins
-import pickle
 import ast
-from ast import fix_missing_locations, alias, ImportFrom
-import traceback
+# import logging
+
+# logging.basicConfig(filename='tmp/bar.log',level=logging.DEBUG)
 
 # things that have to be defined before importing ayton.execute :(
 # singleton
@@ -33,75 +32,23 @@ runner= None
 from ayrton.castt import CrazyASTTransformer
 from ayrton.execute import o, Command, Capture, CommandFailed
 
-__version__= '0.4.1'
+__version__= '0.4.2'
 
-class Environment (object):
-    # this class handles not only environment variables
-    # but also locals, globals and python and ayrton builtins
-    # the latter are only other python functions that are 'promoted' to builtins
+class Ayrton (object):
     def __init__ (self, globals=None, locals=None, **kwargs):
-        super ().__init__ ()
-
         if globals is None:
             self.globals= {}
         else:
             self.globals= globals
+        polute (self.globals, kwargs)
 
         if locals is None:
-            self.locals= {}
+            self.locals= None
         else:
             self.locals= locals
 
-        self.python_builtins= builtins.__dict__.copy ()
-        self.ayrton_builtins= {}
-        polute (self.ayrton_builtins)
-        self.os_environ= os.environ.copy ()
-
-        # now polute the locals with kwargs
-        for k, v in kwargs.items ():
-            # BUG: this sucks
-            if k=='argv':
-                self.ayrton_builtins['argv']= v
-            else:
-                self.locals[k]= v
-
-    def __getitem__ (self, k):
-        strikes= 0
-        for d in (self.locals, self.globals, self.os_environ,
-                  self.python_builtins, self.ayrton_builtins):
-            try:
-                ans= d[k]
-                # found, don't search anymore (just in case you could find it
-                # somewhere else)
-                break
-            except KeyError:
-                strikes+= 1
-
-        if strikes==5:
-            # the name was not found in any of the dicts
-            # create a command for it
-            # ans= Command (k)
-            # print (k)
-            raise KeyError (k)
-
-        return ans
-
-    def __setitem__ (self, k, v):
-        self.locals[k]= v
-
-    def __delitem__ (self, k):
-        del self.locals[k]
-
-    def __iter__ (self):
-        return self.locals.__iter__ ()
-
-    def __str__ (self):
-        return str ([ self.globals, self.locals, self.os_environ ])
-
-class Ayrton (object):
-    def __init__ (self, globals=None, locals=None, **kwargs):
-        self.environ= Environment (globals, locals, **kwargs)
         self.options= {}
+        self.pending_children= []
 
     def run_file (self, file):
         # it's a pity that parse() does not accept a file as input
@@ -110,15 +57,7 @@ class Ayrton (object):
 
     def run_script (self, script, file_name):
         tree= ast.parse (script)
-        # ImportFrom(module='bar', names=[alias(name='baz', asname=None)], level=0)
-        node= ImportFrom (module='ayrton.execute',
-                          names=[alias (name='Command', asname=None)],
-                          level=0)
-        node.lineno= 0
-        node.col_offset= 0
-        ast.fix_missing_locations (node)
-        tree.body.insert (0, node)
-        tree= CrazyASTTransformer(self.environ).visit (tree)
+        tree= CrazyASTTransformer (self.globals).visit (tree)
 
         self.run_tree (tree, file_name)
 
@@ -126,24 +65,34 @@ class Ayrton (object):
         self.run_code (compile (tree, file_name, 'exec'))
 
     def run_code (self, code):
-        exec (code, self.environ.globals, self.environ)
+        exec (code, self.globals, self.locals)
 
-def polute (d):
+    def wait_for_pending_children (self):
+        for i in range (len (self.pending_children)):
+            child= self.pending_children.pop (0)
+            child.wait ()
+
+def polute (d, more):
+    # TODO: weed out some stuff (copyright, etc)
+    d.update (__builtins__)
+    d.update (os.environ)
+
     # these functions will be loaded from each module and put in the globals
     # tuples (src, dst) renames function src to dst
     builtins= {
         'os': [ ('getcwd', 'pwd'), 'uname', 'listdir', ],
         'os.path': [ 'abspath', 'basename', 'commonprefix', 'dirname',  ],
         'time': [ 'sleep', ],
-        'sys': [ 'exit' ], # argv is poluted from the CLI options. see Environment()
+        'sys': [ 'exit', 'argv' ],
 
         'ayrton.file_test': [ '_a', '_b', '_c', '_d', '_e', '_f', '_g', '_h',
                               '_k', '_p', '_r', '_s', '_u', '_w', '_x', '_L',
                               '_N', '_S', '_nt', '_ot' ],
         'ayrton.expansion': [ 'bash', ],
-        'ayrton.functions': [ 'cd', 'export', 'option', 'remote', 'run',
+        'ayrton.functions': [ 'cd', ('cd', 'chdir'), 'export', 'option', 'remote', 'run',
                                'shift', 'source', 'unset', ],
-        'ayrton.execute': [ 'o', 'Capture', 'CommandFailed', 'CommandNotFound', ],
+        'ayrton.execute': [ 'o', 'Capture', 'CommandFailed', 'CommandNotFound',
+                            'Pipe', 'Command'],
         }
 
     for module, functions in builtins.items ():
@@ -160,6 +109,8 @@ def polute (d):
     # now the IO files
     for std in ('stdin', 'stdout', 'stderr'):
         d[std]= getattr (sys, std).buffer
+
+    d.update (more)
 
 def run_tree (tree, globals, locals):
     global runner