+++ ChangeLog.rst
@@ -1,13 +1,31 @@
-ayrton (0.4) UNRELEASED; urgency=low
+ayrton (0.4.2) unstable; urgency=low
+
+  * _bg allows running a command in the background.
+  * _fails allows a Command to fail even when option('-e') is on. 
+  * Try program_name as program-name if the first failed the path lookup.
+  * Convert all arguments to commands to str().
+  * chdir() is an alias of cd().
+  * Capture is a class, not an arbitrary value.
+  * Updated doc.
+  * Fixed globals and local passed to the execution of the script.
+  * Fixed some fd leakage.
+  * Fixed redirection when _out and _err where Capture.
+  * Fixed keyword handling while doing our black magic.
+  * More, better unit tests!
+
+ -- Marcos Dione <mdione@grulic.org.ar>  Wed, 14 Jan 2015 21:58:28 +0100
+
+ayrton (0.4) unstable; urgency=low
 
   * >= can redirect stederr to stdout.
   * o(option=argument) can be used to declare keyword params among/before 
     positional ones.
   * bash() now returns a single string if there is only one result.
-  * Slightly better error reporting: don't print the part of the stacktrace
-    that belongs to `ayrton` itself.
+  * Slightly better error reporting: don't print a part of the stacktrace
+    that belongs to `ayrton` itself. There is still more to do.
+  * No longer depends on `sh`.
 
- -- Marcos Dione <mdione@diablo>  Thu, 05 Dec 2013 13:33:07 +0100
+ -- Marcos Dione <mdione@grulic.org.ar>  Tue, 14 Jan 2014 21:35:13 +0100
 
 ayrton (0.3) unstable; urgency=low